Partial fix for Issue 88: Add Linux amd64 support to JOGL
The stub_includes headers and GlueGen config files have been changed
to make X identifiers and similar values appear to Java to be 64-bit
values on all platforms, rather than changing between 32-bit and
64-bit values depending on the architecture. This allows the shared
Java code in the JOGL X11 implementation to work on both 32- and
64-bit architectures, which is an important step toward having
complete 64-bit support in JOGL. Compared to the suggested patch,
there is no longer a need to duplicate the sources in
net.java.games.jogl.impl.x11 because the Java sources are now 64-bit
clean. The change to use dlsym on Linux/AMD64 instead of
glXGetProcAddressARB has been conditionalized under a run-time flag.
Most of the build.xml and all of the GlueGen changes were incorporated
unchanged. Thanks to the original authors of the patch for this work.

At this point "ant linux.amd64" should build a working JOGL. However,
it can not yet be incorporated into the distribution's platform-
independent jogl.jar; in order to do this, the generation of the
StructAccessors for the JAWT and some other X11 data structures by
GlueGen will need to be conditionalized depending on the platform so
that the 32-bit versions go in one directory, the 64-bit versions in
another, and a run-time check is done to determine which version to
instantiate. There are also a few remaining cleanups to generated glue
code such as net.java.games.jogl.impl.x11.GLX.XGetVisualInfo(), which
needs to be made platform-independent by calling the
XVisualInfo.size() accessor rather than hardwiring in the size of the
data structure (although the real solution will probably need to be
more complicated since there will probably be two XVisualInfo
implementations in the jar file).


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@190 232f8b59-042b-4e1e-8c03-345bb8c30851