Fixed Issue 99: intptr_t undefined compiling with VS .NET 2003 VC++
Applied patch from submitter. I am not convinced it is correct because
we definitely needed that #if clause in order to make JOGL compile
with VC7. However, since I don't have a VC7 installation to test with
and since the patch reportedly makes it work with Visual Studio 2003,
I'll assume it's correct.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@198 232f8b59-042b-4e1e-8c03-345bb8c30851