Partial fix for Issue 140: glu.gluBuild2DMipmaps crash
This was a bug in the SGI GLU sample implementation and was reproduced
by converting the attached test program to C and compiling in the code
from the sample implementation. This bug was already fixed in the Mesa
sources as bug 2510 and the patch was found at
http://cvs.freedesktop.org/mesa/Mesa/src/glu/sgi/libutil/mipmap.c?rev=1.6&view=log
by Googling for the terms "mipmap.c last row". The fix is to add a
test-and-clamp for the highest row read from the texture.

There is still a problem with visual artifacts when scaling
non-power-of-two textures that this test case reproduces and that is
continuing to be investigated.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@234 232f8b59-042b-4e1e-8c03-345bb8c30851