Fixed Issue 148: Swing Component Shadow on GLJPanel
The test case specifies alpha bits in the GLCapabilities, which caused
the GLJPanel to create a TYPE_INT_ARGB BufferedImage to store the
results of the readback of the frame buffer for rendering into the
GLJPanel. This appears to not be the correct thing to do, because it
causes this BufferedImage to be alpha-blended with whatever other
rendering is performed into Java2D's backing store. Therefore when the
button was redrawn, its image would show up in this backing store
alpha-blended with the OpenGL rendering results since both were
redrawn simultaneously. I believe the correct solution is to never
create a BufferedImage with an alpha component for the framebuffer
readback even if alpha bits are specified in the GLCapabilities. This
is what has been implemented in the JOGL source base.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@258 232f8b59-042b-4e1e-8c03-345bb8c30851