This putback adds array offsets to the public JOGL API.  The offsets are respected and used properly in all of the public and private functions.  The changes are in gluegen, so that the code is generated properly.  And also throughout the parts of the jogl code that are not gluegen-generated.
For the internally generated implementation methods, a "1" is added to
the method names.  So as to not overload the public API.  This is similar
to what is already done with Buffer APIs, which have a "0" added
internally.  I used a "1" instead of a "0" to avoid any collisions of the
signatures, which could happen if a null object was sent down for the
Array (e.g., the wrong method would get called).

This should be a suitable foundation for the implementation to add the
ability to wrap arrays in Buffers, which we plan to add to the implementation
soon.

These changes will cause all existing JOGL programs to break, although
adapting them is pretty easy.  We will later be putting back changed examples
and utilities that incorporate these new APIs.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/branches/JSR-231@313 232f8b59-042b-4e1e-8c03-345bb8c30851