This putback may break Windows/Mac build temporarily for a few hours until we synchronize with builds on all of our target platforms as we are rushing this a bit so that we can properly reorganize the workspace shortly, and with such fundamental changes, this is considered a "risky" putback. So advising to hold off on bringover a few hours for non-X11 targets.
This is the majority of changes required for Native I/O IndirectBuffer
support for JSR 231.  The changes are not yet complete, and Indirect
Buffers (buffers backed by an array) should not yet be used (although
what is being putback here has been tested and is working).

IndirectBuffer support is being added in JSR 231 since this will
be the method often used to send down arrays to the public API,
by "wrapping" them, enabling us to eliminate many of the API
"explosions" that happen in the older, JOGL approach, enabling
the number of public APIs in JSR 231 to be greatly reduced, and
map more cleanly to native OpenGL.  Soon we
will be removing the old API explosions and Indirect Buffers will
be the only way to send down an array to many APIs, but at this
point there are no public API changes putback for this.  We will
be providing code samples.  The changes to any program source
code are *very* straightforward.

Offsets into Native I/O buffers and offsets within the arrays that
back them are also supported.  We will also provide code examples
for this.  So, for example, you do not need to rewrap an array when
you want to send down its elements but starting at somewhere besides
the zeroth index.

This putback adds the support for all methods not accessed internally by
function pointers (dispatch methods in the expanded code).  The
support for function pointers (mostly APIs that were later additions
to OpenGL, mostly things that are or originally were extensions) will
come in a subsequent putback.  The native code for the dispatch methods
is mostly generated, however (just never called currently).  The Java
part for dispatch methods is not complete yet, and is not yet generated.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/branches/JSR-231@352 232f8b59-042b-4e1e-8c03-345bb8c30851