Initial integration of JOGL with the Java2D OpenGL pipeline in the Sun JDK. GLJPanel has a new rendering path which captures the Java2D back buffer via GLDrawableFactory.createExternalGLDrawable()/createExternalGLContext() and creates a new context on it for performing JOGL rendering. The new path is enabled by default in Mustang build 51 and later when the Java2D/OpenGL pipeline is turned on via -Dsun.java2d.opengl=true, and yields huge speedups relative to the previous pbuffer-based GLJPanel implementation. GLJPanel in the new configuration is nearly as fast as the GLCanvas, and still provides 100% correct Swing integration. No public API changes were required to JOGL in order to implement this integration with one exception in GLJPanel.isOGLPipelineEnabled(): see below.
Issues still remain, such as the desire to run all (not just some) OpenGL
work on the Java2D-internal Queue Flusher Thread to maintain single-threaded
behavior of the library. Currently GLCanvas's OpenGL work is run on the EDT
while GLJPanel's and GLPbuffer's OpenGL work is run on the QFT. Significant
restructuring to the GLCanvas will be required to achieve this goal. There is
some lag of mouse events with demos which are more expensive to render which
needs to be investigated. It will not be possible to implement calling of
display() on other non-pbuffer GLAutoDrawables from within the
GLEventListener's display() callback in this model. There are other issues as
well.

GLJPanel falls back to using a pbuffer internally when the Java2D back
buffer's pixel format is not suitable for rendering the JOGL demos (for
example, the InfiniteShadowVolumes demo, which requires stencil bits).

Added GLJPanel.isOGLPipelineEnabled() which applications can use to decide
whether to clear the color buffer or leave prior Java2D rendering results in
place for non-opaque GLJPanels. The JGears demo has been modified to query
that flag to maintain the gradient behind it while rendering the gears
directly into the Java2D back buffer.

GLCanvas uses the pixel format specified by Java2D if one is already set. It
turns out that when the Java2D OpenGL pipeline is enabled, one is always set
by default even though no Java2D rendering is performed into it, because the
AWT clears the background of Canvases by default in certain situations. This
can be disabled (only globally, unfortunately) by specifying
-Dsun.awt.noerasebackground=true.

Fixed issues with HDR and HWShadowmapsSimple demos where they needed to pay
attention to the x and y coordinates passed down in the reshape callback in
order to set up the correct OpenGL viewport. Fixed minor shutdown-related
issues with demos using gleem's ManipManager.

Tested (on NVidia hardware only so far) on Windows and X11 platforms with
Mustang build 51.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/branches/JSR-231@368 232f8b59-042b-4e1e-8c03-345bb8c30851