Issue number: Obtained from: Submitted by:   Travis Reviewed by:
These are all documentation changes in preparation for the Proposed Final
Draft submission.

Carefully separated out references to the implementation from the
specification.  Did not delete any comments, just made sure that it was
clear when it was the Reference Implementation being referred to.

Add note to Spec Overview about version information.  We now include
version information in the manifest of the jar file and we require
this of all implementations.  This is currently done automatically
as part of the build process and is easy to access at runtime with
the java.lang.Package APIs.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@592 232f8b59-042b-4e1e-8c03-345bb8c30851