Issue number: Obtained from: Submitted by:   Travis Reviewed by:
We are modifying the versioning for non-release builds.  Now the timestamp
is part of the implementation version string.  For instance,
"1.0.0-pre-20060215-17:03:32" is an implementation of the 1.0.0 version
of the API built on Feb. 15, 2006, at 5:03:32 PM.  This is how our daily
build version strings will look.  This is how the default build works.

Our official beta builds which are known as "stable" builds will have a
version string like: "1.0.0-beta3", etc.

Our final version of 1.0.0 implementation will become the "release" build
and the version string will be: 1.0.0


There are two easy ways to obtain the version string:
1) use the jogl.verbose properties flag (-Djogl.verbose) to see it at the
   command line
2) use the java.lang.Package API to query it programatically.  See example
   in demos called demos/misc/VersionInfo

We are removing the Version utility class because it was insufficient for
a number of reasons.  It should be removed with this putback.

I am not 100% positive that this does not break the JOGLAppletLauncher class
with the change that I have putback, but I can't test it here in my
environment.  And I am pretty sure it will still work, but we will test it.


 Modified Files:
 	src/classes/com/sun/opengl/util/JOGLAppletLauncher.java
 	make/joglversion make/build.xml make/joglRIversion
 ----------------------------------------------------------------------


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@612 232f8b59-042b-4e1e-8c03-345bb8c30851