Issue number: Obtained from: Submitted by:  Travis Reviewed by:
Add 64 bit Sparcv9 Solaris build.  A concomitant change will be putback
into the gluegen tree shortly for properly detecting this case.

The way that gluegen checks 32 or 64 bit-ness is which java binary is in
your path, 32 bit sparc java or 64 bit sparcv9 java.

In order to build the jogl 64 bit target, you need to have installed the
64 bit OpenGL sparcv9 libraries.

This builds the JOGL binaries in the same place as the 32 bit JOGL .so
library builds binaries, so you can build one or the other with a given
source tree but not both.

The jogl.jar generated by either the 32 or 64 bit target it identical,
however.

I have not added the 64 bit target to the dist build, that is coming soon.
As well, it will be added to the nightly builds soon along with Linux AMD64
and Solaris AMD64.


 Modified Files:
 	make/build.xml
 ----------------------------------------------------------------------


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@736 232f8b59-042b-4e1e-8c03-345bb8c30851