Fixed Issue 220: Gears demo throws exceptions with -Dsun.java2d.opengl=true
On non-X11 platforms, removed workaround in Threading class which
treats both the EDT and the QFT as viable threads upon which to
execute OpenGL work. Deadlock issues remain on X11 threads when
causing OpenGL contexts associated with heavyweight widgets to be made
current on the QFT. On Windows the QFT is now used for all OpenGL work
when the Java2D/JOGL bridge is active.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@786 232f8b59-042b-4e1e-8c03-345bb8c30851