Issue number: Obtained from: Submitted by:  Travis Reviewed by:
Adding support for 64 bit Solaris on AMD64.  Make sure the 64 bit JVM is
in your path and it will build the 64 instead of the 32 target.

We are still having problems with this variant running properly in many
cases and I will file an issue.

I will putback the accompanying code in gluegen shortly.


 Modified Files:
 	make/build.xml


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@809 232f8b59-042b-4e1e-8c03-345bb8c30851