Issue number: Obtained from: Submitted by:   Travis Reviewed by:
Sun is adding an alternative license term to the SGI Free B license which
is permitted under that license.  We are removing a couple of provisions
from the SGI Free B license to make the license more liberal for our end
users of JOGL.  So these changes are just communication of that.  There are
no actual source code changes, just changes to the license file and the
license text in each source code file.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@856 232f8b59-042b-4e1e-8c03-345bb8c30851