Fixed Issue 262: DRIHack problems with ATI and possibly other drivers
Added support to GlueGen's NativeLibrary class for relative library
path names which do not require expansion; for example, "libGL.so.1".
Changed the DRIHack to first attempt to open libGL.so.1 instead of
"GL", which expanded to "libGL.so". For some reason, this causes ATI's
drivers to be picked up properly. Tested with various JOGL demos.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@1093 232f8b59-042b-4e1e-8c03-345bb8c30851