Fixed Issue 275: Rendering bugs in TextRenderer
Fixed problem where during resizing of backing store for TextRenderer
we were neglecting to unbind the old texture and bind the new one.
This caused attempts to render with an invalid texture, which was
probably the cause of crashes and definitely the cause of rendering
artifacts. Fixed by calling endRendering() / beginRendering()
appropriately on the underlying TextureRenderer during resizing of the
backing store of the TextRenderer.


git-svn-id: file:///usr/local/projects/SUN/JOGL/git-svn/svn-server-sync/jogl/trunk@1148 232f8b59-042b-4e1e-8c03-345bb8c30851